separate Rendering from Sim: start with BeamLaserProjectile
The idea is to register drawer instances in some central point
(for projectiles in ProjectileDrawer), have two drawer instances per
projectile type (the default, 3D OpenGL one and a mini-map one).
In a first step, the Draw*() methods are keeps as is in the Sim code,
and are re-routed to the Drawer implementations, to be able to do the change
gradually.
Once all drawing code is moved to Drawer implementations, the remaining
drawing stuff will be removed from the Sim part of projectiles.